further tidying of stone example
