+++ projects/stone_cx/CMakeLists.txt
@@ -17,14 +17,8 @@ elseif(${TARGET} STREQUAL simulatorReplay)
     set(SOURCES simulatorReplay.cc visualizationCommon.cc)
     set(BOB_MODULES common)
 elseif(${TARGET} STREQUAL robot)
-    set(SOURCES robot.cc robotCommon.cc)
-    set(BOB_MODULES common hid imgproc robots video)
-elseif(${TARGET} STREQUAL robotDeadReckon)
-    set(SOURCES robotDeadReckon.cc visualizationCommon.cc)
-    set(BOB_MODULES common hid imgproc net os robots video)
-elseif(${TARGET} STREQUAL robotVicon)
-    set(SOURCES robotVicon.cc robotCommon.cc)
-    set(BOB_MODULES common hid imgproc robots video vicon)
+    set(SOURCES robot.cc visualizationCommon.cc)
+    set(BOB_MODULES common hid imgproc net os robots video vicon)
 else()
     message(FATAL_ERROR "Bad target specified")
 endif()
