--- projects/stone_cx/robotCommon.h
@@ -1,14 +1,0 @@
-#pragma once
-
-namespace BoBRobotics {
-namespace Robots {
-// Forward declarations
-class Tank;
-class ROBOT_TYPE;
-}
-
-namespace StoneCX {
-// Functions
-float driveMotorFromCPU1(BoBRobotics::Robots::Tank &motor, bool display = false);
-}
-}
